refactor(dashboard): rename realunit-tracing files to log-tracing prefix
Restructure file naming ahead of adding generic log views:
- src/screens/dashboard-realunit-tracing.screen.tsx
    -> src/screens/dashboard-log-tracing-realunit.screen.tsx
- src/hooks/realunit-tracing.hook.ts -> src/hooks/log-tracing.hook.ts
- src/components/dashboard/realunit-trace-time-chart.tsx
    -> src/components/dashboard/log-trace-time-chart.tsx

Identifier renames:
- useRealunitTracing -> useLogTracing
- RealUnitTraceTimeChart -> LogTraceTimeChart
- DashboardRealunitTracingScreen -> DashboardLogTracingRealunitScreen

getRealunitTraces stays — it remains the RealUnit-specific fetcher.
